requirements.txt
python-dotenv
duckdb
pandas
numpy
TA-Lib
alpaca-py
requests
schedule
pytz
pytest
